Move approval list and edit view to separate components
